fixup: Split apart sui-framework-tests
